🐛 align share limiter with relocated HTTP boundary
